Normalize files from upstream additions
