+++ components/colors.ts
