+++ pages/_app.tsx
@@ -2,8 +2,8 @@ import { CSSObject, Global } from '@emotion/core'
 import React from 'react'
 import App, { Container } from 'next/app'
 import Head from 'next/head'
-import Footer from '../wtf/Footer'
-import Navbar from '../wtf/Navbar'
+import Footer from '../components/Footer'
+import Navbar from '../components/Navbar'
 
 const STYLES: CSSObject = {
   '*, *::before, *::after': {
@@ -34,7 +34,7 @@ const STYLES: CSSObject = {
   },
 }
 
-export default class MyApp extends App {
+export default class Application extends App {
   render() {
     const { Component, pageProps } = this.props
     return (
