--- posts/draft.md
@@ -1,1 +1,0 @@
-# This is a Draft
