--- posts/examples/tubs.md
@@ -1,11 +1,0 @@
----
-author: MK
-date: 06/07/2006
-image: https://source.unsplash.com/daily?tubs
-summary: Hey there. Please make the text length here so it's about 80 characters...
-title: Tubs.md
----
-
-# Tubs.md
-
-More text
